Add weekday terms validation to EntityModel
Additional validation has been added to EntityModel that allows the inclusion of days of the week in term models. A new class, AddedWeekdayTermsValidationMixin, has been created to implement this functionality. With this update, developers now have the ability to validate specific weekdays as terms, further enhancing the model's data structuring capabilities.