Refactor query method in entity model and update custom_group_by method
The primary action in this commit is the refactored query method in entity model to modify the construction of result queryset, allowing a more efficient handling of data grouping. In addition, it includes an update in the 'custom_group_by' method in the SQLCompiler class. Moreover, debug prints were added to aid in debugging for future adjustments.