Add 'select_across' field to forms and implement pagination
This commit adds the 'select_across' field to multiple forms as a hidden input. It is to store the value of 'select_across' from the post request. Also, it introduces pagination to queryset processes, which allows displaying only a portion of all the items in a list, reducing the load on the application.